correccion y orden de ultimo momento
