+++ index.html
@@ -505,6 +505,6 @@
       integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
       crossorigin="anonymous"
     ></script>
-    <script src="assets/js/script.js"></script>
+    <script defer type="module" src="assets/js/index.js"></script>
   </body>
 </html>
